requirements.txt
alembic==1.10.2
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
ecdsa==0.18.0
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
pyasn1==0.4.8
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2022.7.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
validators==0.20.0
Werkzeug==2.2.3
